requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.2
pytest==8.3.3
email-validator==2.2.0
openpyxl==3.1.5
reportlab==4.2.2
httpx==0.27.2
python-jose==3.3.0
psycopg2-binary==2.9.10
